1647: add store links to buildConfigs, extract app store link container to separate component to reuse it
