+++ frontend/build-configs/types.ts
@@ -109,12 +109,14 @@ export type AndroidBuildConfigType = CommonBuildConfigType & {
         excludeLocationPlayServices: boolean
         excludeX86: boolean
     }
+    appStoreLink: string
 }
 
 export type iOSBuildConfigType = CommonBuildConfigType & {
     // iOS application identifier.
     bundleIdentifier: string
     provisioningProfileSpecifier: string
+    appStoreLink: string
 }
 
 export default BuildConfigType
